Made answer_question work with new architecture
