+++ .gitignore
@@ -155,7 +155,7 @@ dmypy.json
 cython_debug/
 
 # local development
-ollama/
+dev_vector_db/
 
 # PyCharm
 #  JetBrains specific template is maintained in a separate JetBrains.gitignore that can
