+++ src/parsers/__init__.py
@@ -1,0 +1,1 @@
+from .json_parser import get_json_parser
