+++ src/prompts/__init__.py
@@ -1,0 +1,1 @@
+from .get_inconcistencies_prompt import get_inconcistencies_prompt
